# swagger.yaml
swagger: '2.0'
info:
  version: "0.0.4"
  title: Wave Application
host: localhost:9600
basePath: /
...
schemes:
  - http
paths:
...
  /users/leaders?count=${count}&page=${page}:
     get:
      tags:
      - user
      summary: "Get all users with their scores."
      description: ""
      operationId: leaders
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        406:
          description: Not Acceptable
        400:
          description: Bad Request
        200:
          description: OK
          schema:
            $ref: "#/definitions/Leaderboard"
        500:
          description: Internal Server Error
...
definitions:
...
  UserScore:
    type: object
    required:
      - username
      - score
    properties:
      username:
        type: string
        example: <username>
      score:
        type: integer
        example: 0
  Leaderboard:
    type: object
    properties:
      users:
        type: array
        items:
          $ref: "#/definitions/UserScore"
      total:
        type: integer
        minimum: 0
        example: 0
...